Split order mutations files into file per mutation
