+++ saleor/graphql/order/mutations/draft_order_create.py
@@ -5,41 +5,25 @@ from django.core.exceptions import ValidationError
 from django.db import transaction
 from graphene.types import InputObjectType
 
-from ....account.models import User
 from ....checkout import AddressType
-from ....core.exceptions import InsufficientStock
 from ....core.permissions import OrderPermissions
-from ....core.taxes import TaxError, zero_taxed_money
+from ....core.taxes import TaxError
 from ....core.tracing import traced_atomic_transaction
 from ....core.utils.url import validate_storefront_url
 from ....order import OrderOrigin, OrderStatus, events, models
-from ....order.actions import order_created
 from ....order.error_codes import OrderErrorCode
-from ....order.fetch import OrderInfo, OrderLineInfo
-from ....order.search import (
-    prepare_order_search_document_value,
-    update_order_search_document,
-)
-from ....order.utils import (
-    add_variant_to_order,
-    get_order_country,
-    recalculate_order,
-    update_order_prices,
-)
-from ....warehouse.management import allocate_preorders, allocate_stocks
-from ....warehouse.reservations import is_reservation_enabled
+from ....order.search import update_order_search_document
+from ....order.utils import add_variant_to_order, recalculate_order, update_order_prices
 from ...account.i18n import I18nMixin
 from ...account.types import AddressInput
 from ...channel.types import Channel
-from ...core.mutations import BaseMutation, ModelDeleteMutation, ModelMutation
+from ...core.mutations import ModelMutation
 from ...core.scalars import PositiveDecimal
 from ...core.types import NonNullList, OrderError
 from ...product.types import ProductVariant
 from ...shipping.utils import get_shipping_model_by_object_id
 from ..types import Order, OrderLine
 from ..utils import (
-    prepare_insufficient_stock_order_validation_errors,
-    validate_draft_order,
     validate_product_is_published_in_channel,
     validate_variant_channel_listings,
 )
@@ -346,184 +330,3 @@ class DraftOrderCreate(ModelMutation, I18nMixin):
         # Post-process the results
         recalculate_order(instance)
         update_order_search_document(instance)
-
-
-class DraftOrderUpdate(DraftOrderCreate):
-    class Arguments:
-        id = graphene.ID(required=True, description="ID of a draft order to update.")
-        input = DraftOrderInput(
-            required=True, description="Fields required to update an order."
-        )
-
-    class Meta:
-        description = "Updates a draft order."
-        model = models.Order
-        object_type = Order
-        permissions = (OrderPermissions.MANAGE_ORDERS,)
-        error_type_class = OrderError
-        error_type_field = "order_errors"
-
-    @classmethod
-    def get_instance(cls, info, **data):
-        instance = super().get_instance(
-            info, qs=models.Order.objects.prefetch_related("lines"), **data
-        )
-        if instance.status != OrderStatus.DRAFT:
-            raise ValidationError(
-                {
-                    "id": ValidationError(
-                        "Provided order id belongs to non-draft order. "
-                        "Use `orderUpdate` mutation instead.",
-                        code=OrderErrorCode.INVALID,
-                    )
-                }
-            )
-        return instance
-
-    @classmethod
-    def save(cls, info, instance, cleaned_input):
-        return cls._save_draft_order(info, instance, cleaned_input, new_instance=False)
-
-
-class DraftOrderDelete(ModelDeleteMutation):
-    class Arguments:
-        id = graphene.ID(required=True, description="ID of a draft order to delete.")
-
-    class Meta:
-        description = "Deletes a draft order."
-        model = models.Order
-        object_type = Order
-        permissions = (OrderPermissions.MANAGE_ORDERS,)
-        error_type_class = OrderError
-        error_type_field = "order_errors"
-
-    @classmethod
-    def clean_instance(cls, info, instance):
-        if instance.status != OrderStatus.DRAFT:
-            raise ValidationError(
-                {
-                    "id": ValidationError(
-                        "Provided order id belongs to non-draft order.",
-                        code=OrderErrorCode.INVALID,
-                    )
-                }
-            )
-
-    @classmethod
-    @traced_atomic_transaction()
-    def perform_mutation(cls, _root, info, **data):
-        order = cls.get_instance(info, **data)
-        response = super().perform_mutation(_root, info, **data)
-        transaction.on_commit(lambda: info.context.plugins.draft_order_deleted(order))
-        return response
-
-
-class DraftOrderComplete(BaseMutation):
-    order = graphene.Field(Order, description="Completed order.")
-
-    class Arguments:
-        id = graphene.ID(
-            required=True, description="ID of the order that will be completed."
-        )
-
-    class Meta:
-        description = "Completes creating an order."
-        permissions = (OrderPermissions.MANAGE_ORDERS,)
-        error_type_class = OrderError
-        error_type_field = "order_errors"
-
-    @classmethod
-    def update_user_fields(cls, order):
-        if order.user:
-            order.user_email = order.user.email
-        elif order.user_email:
-            try:
-                order.user = User.objects.get(email=order.user_email)
-            except User.DoesNotExist:
-                order.user = None
-
-    @classmethod
-    def validate_order(cls, order):
-        if not order.is_draft():
-            raise ValidationError(
-                {
-                    "id": ValidationError(
-                        "The order is not draft.", code=OrderErrorCode.INVALID.value
-                    )
-                }
-            )
-
-    @classmethod
-    def perform_mutation(cls, _root, info, id):
-        manager = info.context.plugins
-        order = cls.get_node_or_error(
-            info,
-            id,
-            only_type=Order,
-            qs=models.Order.objects.prefetch_related("lines__variant"),
-        )
-        cls.validate_order(order)
-
-        country = get_order_country(order)
-        validate_draft_order(order, country, info.context.plugins)
-        cls.update_user_fields(order)
-        order.status = OrderStatus.UNFULFILLED
-
-        if not order.is_shipping_required():
-            order.shipping_method_name = None
-            order.shipping_price = zero_taxed_money(order.currency)
-            if order.shipping_address:
-                order.shipping_address.delete()
-                order.shipping_address = None
-
-        order.search_document = prepare_order_search_document_value(order)
-        order.save()
-
-        channel = order.channel
-        channel_slug = channel.slug
-        order_lines_info = []
-        for line in order.lines.all():
-            if line.variant.track_inventory or line.variant.is_preorder_active():
-                line_data = OrderLineInfo(
-                    line=line, quantity=line.quantity, variant=line.variant
-                )
-                order_lines_info.append(line_data)
-                try:
-                    with traced_atomic_transaction():
-                        allocate_stocks(
-                            [line_data],
-                            country,
-                            channel_slug,
-                            manager,
-                            check_reservations=is_reservation_enabled(
-                                info.context.site.settings
-                            ),
-                        )
-                        allocate_preorders(
-                            [line_data],
-                            channel_slug,
-                            check_reservations=is_reservation_enabled(
-                                info.context.site.settings
-                            ),
-                        )
-                except InsufficientStock as exc:
-                    errors = prepare_insufficient_stock_order_validation_errors(exc)
-                    raise ValidationError({"lines": errors})
-
-        order_info = OrderInfo(
-            order=order,
-            customer_email=order.get_customer_email(),
-            channel=channel,
-            payment=order.get_last_payment(),
-            lines_data=order_lines_info,
-        )
-
-        order_created(
-            order_info=order_info,
-            user=info.context.user,
-            app=info.context.app,
-            manager=info.context.plugins,
-            from_draft=True,
-        )
-
-        return DraftOrderComplete(order=order)
