+++ saleor/graphql/order/mutations/order_add_note.py
@@ -1,0 +1,70 @@
+import graphene
+from django.core.exceptions import ValidationError
+from django.db import transaction
+
+from ....core.permissions import OrderPermissions
+from ....core.tracing import traced_atomic_transaction
+from ....order import events
+from ....order.error_codes import OrderErrorCode
+from ...core.mutations import BaseMutation
+from ...core.types import OrderError
+from ...core.utils import validate_required_string_field
+from ..types import Order, OrderEvent
+from .utils import get_webhook_handler_by_order_status
+
+
+class OrderAddNoteInput(graphene.InputObjectType):
+    message = graphene.String(
+        description="Note message.", name="message", required=True
+    )
+
+
+class OrderAddNote(BaseMutation):
+    order = graphene.Field(Order, description="Order with the note added.")
+    event = graphene.Field(OrderEvent, description="Order note created.")
+
+    class Arguments:
+        id = graphene.ID(
+            required=True,
+            description="ID of the order to add a note for.",
+            name="order",
+        )
+        input = OrderAddNoteInput(
+            required=True, description="Fields required to create a note for the order."
+        )
+
+    class Meta:
+        description = "Adds note to the order."
+        permissions = (OrderPermissions.MANAGE_ORDERS,)
+        error_type_class = OrderError
+        error_type_field = "order_errors"
+
+    @classmethod
+    def clean_input(cls, _info, _instance, data):
+        try:
+            cleaned_input = validate_required_string_field(data["input"], "message")
+        except ValidationError:
+            raise ValidationError(
+                {
+                    "message": ValidationError(
+                        "Message can't be empty.",
+                        code=OrderErrorCode.REQUIRED,
+                    )
+                }
+            )
+        return cleaned_input
+
+    @classmethod
+    @traced_atomic_transaction()
+    def perform_mutation(cls, _root, info, **data):
+        order = cls.get_node_or_error(info, data.get("id"), only_type=Order)
+        cleaned_input = cls.clean_input(info, order, data)
+        event = events.order_note_added_event(
+            order=order,
+            user=info.context.user,
+            app=info.context.app,
+            message=cleaned_input["message"],
+        )
+        func = get_webhook_handler_by_order_status(order.status, info)
+        transaction.on_commit(lambda: func(order))
+        return OrderAddNote(order=order, event=event)
