+++ saleor/graphql/order/mutations/order_mark_as_paid.py
@@ -1,0 +1,52 @@
+import graphene
+from django.core.exceptions import ValidationError
+
+from ....core.permissions import OrderPermissions
+from ....order.actions import clean_mark_order_as_paid, mark_order_as_paid
+from ....order.error_codes import OrderErrorCode
+from ....order.search import update_order_search_document
+from ...core.mutations import BaseMutation
+from ...core.types import OrderError
+from ..types import Order
+from .utils import try_payment_action
+
+
+class OrderMarkAsPaid(BaseMutation):
+    order = graphene.Field(Order, description="Order marked as paid.")
+
+    class Arguments:
+        id = graphene.ID(required=True, description="ID of the order to mark paid.")
+        transaction_reference = graphene.String(
+            required=False, description="The external transaction reference."
+        )
+
+    class Meta:
+        description = "Mark order as manually paid."
+        permissions = (OrderPermissions.MANAGE_ORDERS,)
+        error_type_class = OrderError
+        error_type_field = "order_errors"
+
+    @classmethod
+    def clean_billing_address(cls, instance):
+        if not instance.billing_address:
+            raise ValidationError(
+                "Order billing address is required to mark order as paid.",
+                code=OrderErrorCode.BILLING_ADDRESS_NOT_SET,
+            )
+
+    @classmethod
+    def perform_mutation(cls, _root, info, **data):
+        order = cls.get_node_or_error(info, data.get("id"), only_type=Order)
+        transaction_reference = data.get("transaction_reference")
+        cls.clean_billing_address(order)
+        user = info.context.user
+        app = info.context.app
+        try_payment_action(order, user, app, None, clean_mark_order_as_paid, order)
+
+        mark_order_as_paid(
+            order, user, app, info.context.plugins, transaction_reference
+        )
+
+        update_order_search_document(order)
+
+        return OrderMarkAsPaid(order=order)
