+++ saleor/graphql/order/schema.py
@@ -11,41 +11,33 @@ from ..core.utils import from_global_id_or_error
 from .bulk_mutations.draft_orders import DraftOrderBulkDelete, DraftOrderLinesBulkDelete
 from .bulk_mutations.orders import OrderBulkCancel
 from .filters import DraftOrderFilter, OrderFilter
-from .mutations.discount_order import (
-    OrderDiscountAdd,
-    OrderDiscountDelete,
-    OrderDiscountUpdate,
-    OrderLineDiscountRemove,
-    OrderLineDiscountUpdate,
-)
-from .mutations.draft_orders import (
-    DraftOrderComplete,
-    DraftOrderCreate,
-    DraftOrderDelete,
-    DraftOrderUpdate,
-)
-from .mutations.fulfillments import (
-    FulfillmentApprove,
-    FulfillmentCancel,
-    FulfillmentRefundProducts,
-    FulfillmentReturnProducts,
-    FulfillmentUpdateTracking,
-    OrderFulfill,
-)
-from .mutations.orders import (
-    OrderAddNote,
-    OrderCancel,
-    OrderCapture,
-    OrderConfirm,
-    OrderLineDelete,
-    OrderLinesCreate,
-    OrderLineUpdate,
-    OrderMarkAsPaid,
-    OrderRefund,
-    OrderUpdate,
-    OrderUpdateShipping,
-    OrderVoid,
-)
+from .mutations.draft_order_complete import DraftOrderComplete
+from .mutations.draft_order_create import DraftOrderCreate
+from .mutations.draft_order_delete import DraftOrderDelete
+from .mutations.draft_order_update import DraftOrderUpdate
+from .mutations.fulfillment_approve import FulfillmentApprove
+from .mutations.fulfillment_cancel import FulfillmentCancel
+from .mutations.fulfillment_refund_products import FulfillmentRefundProducts
+from .mutations.fulfillment_return_products import FulfillmentReturnProducts
+from .mutations.fulfillment_update_tracking import FulfillmentUpdateTracking
+from .mutations.order_add_note import OrderAddNote
+from .mutations.order_cancel import OrderCancel
+from .mutations.order_capture import OrderCapture
+from .mutations.order_confirm import OrderConfirm
+from .mutations.order_discount_add import OrderDiscountAdd
+from .mutations.order_discount_delete import OrderDiscountDelete
+from .mutations.order_discount_update import OrderDiscountUpdate
+from .mutations.order_fulfill import OrderFulfill
+from .mutations.order_line_delete import OrderLineDelete
+from .mutations.order_line_discount_remove import OrderLineDiscountRemove
+from .mutations.order_line_discount_update import OrderLineDiscountUpdate
+from .mutations.order_line_update import OrderLineUpdate
+from .mutations.order_lines_create import OrderLinesCreate
+from .mutations.order_mark_as_paid import OrderMarkAsPaid
+from .mutations.order_refund import OrderRefund
+from .mutations.order_update import OrderUpdate
+from .mutations.order_update_shipping import OrderUpdateShipping
+from .mutations.order_void import OrderVoid
 from .resolvers import (
     resolve_draft_orders,
     resolve_homepage_events,
